Code Generalization - Update
Removed the specifics of "ADCP" and "GPS" loops and allowed for the user to input the names of the alternative data sources. Note: these alternative data sources must be formatted as it is indicated in the Post-Hysweep Processing Procedures with only an X, Y, and Z column, saved as a .csv file, and saved under the naming convention of [LakeName]_[datasource].csv in a folder sharing the same name as the alternative data source. For example, GardenCityOId_ADPC.csv in the ADCP subdirectory of the Garden City Old lake directory.